put unfocus button in menu and fix some ui issues
